+++ spec/helpers/src/prepareGame.ts
@@ -4,6 +4,7 @@ import * as g from "@akashic/akashic-engine";
 import * as mockrf from "./MockResourceFactory";
 import { MockGame } from "./MockGame";
 import { PdiUtil } from "../../../lib/PdiUtil";
+import { GameHandlerSet } from "../../../lib/GameHandlerSet";
 
 export enum FixtureGame {
 	/**
@@ -33,9 +34,12 @@ export function prepareGame(param: PrepareGameParameterObject): MockGame {
 	configuration = PdiUtil._resolveConfigurationBasePath(configuration, assetBase);
 
 	var game = new MockGame({
+		engineModule: g,
 		configuration: configuration,
+		handlerSet: new GameHandlerSet({ isSnapshotSaver: false }),
 		resourceFactory: new mockrf.ResourceFactory(),
 		assetBase: assetBase,
+		selfId: param.playerId,
 		player: param.player ? param.player : { id: param.playerId }
 	});
 	return game;