+++ src/Game.ts
@@ -1,23 +1,13 @@
 "use strict";
 import * as g from "@akashic/akashic-engine";
-import * as amf from "@akashic/amflow";
 import StartPointData from "./StartPointData";
 import { StorageFunc } from "./StorageFunc";
+import { GameHandlerSet } from "./GameHandlerSet";
 
-export interface GameParameterObject {
-	configuration: g.GameConfiguration;
-	resourceFactory: g.ResourceFactory;
-	assetBase: string;
+export interface GameParameterObject extends g.GameParameterObject {
 	player: g.Player;
-	isSnapshotSaver?: boolean;
-	operationPluginViewInfo?: g.OperationPluginViewInfo;
 	gameArgs?: any;
 	globalGameArgs?: any;
-}
-
-export interface GameEventFilterFuncs {
-	addFilter: (filter: g.EventFilter, handleEmpty?: boolean) => void;
-	removeFilter: (filter?: g.EventFilter) => void;
 }
 
 /**
@@ -60,12 +50,7 @@ export class Game extends g.Game {
 	abortTrigger: g.Trigger<void>;
 
 	player: g.Player;
-	raiseEventTrigger: g.Trigger<g.Event>;
-	raiseTickTrigger: g.Trigger<g.Event[]>;
-	snapshotTrigger: g.Trigger<amf.StartPoint>;
-	isSnapshotSaver: boolean;
-	_getCurrentTimeFunc: () => number;
-	_eventFilterFuncs: GameEventFilterFuncs;
+	handlerSet: GameHandlerSet;
 	_notifyPassedAgeTable: { [age: number]: boolean };
 	_notifiesTargetTimeReached: boolean;
 	_isSkipAware: boolean;
@@ -73,18 +58,12 @@ export class Game extends g.Game {
 	_globalGameArgs: any;
 
 	constructor(param: GameParameterObject) {
-		super(param.configuration, param.resourceFactory, param.assetBase, param.player.id, param.operationPluginViewInfo);
+		super(param);
 		this.agePassedTrigger = new g.Trigger<number>();
 		this.targetTimeReachedTrigger = new g.Trigger<number>();
 		this.skippingChangedTrigger = new g.Trigger<boolean>();
 		this.abortTrigger = new g.Trigger<void>();
 		this.player = param.player;
-		this.raiseEventTrigger = new g.Trigger<g.Event>();
-		this.raiseTickTrigger = new g.Trigger<g.Event[]>();
-		this.snapshotTrigger = new g.Trigger<amf.StartPoint>();
-		this.isSnapshotSaver = !!param.isSnapshotSaver;
-		this._getCurrentTimeFunc = null;
-		this._eventFilterFuncs = null;
 		this._notifyPassedAgeTable = {};
 		this._notifiesTargetTimeReached = false;
 		this._isSkipAware = false;
@@ -134,22 +113,6 @@ export class Game extends g.Game {
 		return false;
 	}
 
-	/**
-	 * `Game` が利用する時刻取得関数をセットする。
-	 * このメソッドは `Game#_load()` 呼び出しに先行して呼び出されていなければならない。
-	 */
-	setCurrentTimeFunc(fun: () => number): void {
-		this._getCurrentTimeFunc = fun;
-	}
-
-	/**
-	 * `Game` のイベントフィルタ関連実装をセットする。
-	 * このメソッドは `Game#_load()` 呼び出しに先行して呼び出されていなければならない。
-	 */
-	setEventFilterFuncs(funcs: GameEventFilterFuncs): void {
-		this._eventFilterFuncs = funcs;
-	}
-
 	setStorageFunc(funcs: StorageFunc): void {
 		this.storage._registerLoad(funcs.storageGetFunc);
 		this.storage._registerWrite(funcs.storagePutFunc);
@@ -165,52 +128,6 @@ export class Game extends g.Game {
 		this._isSkipAware = aware;
 	}
 
-	getCurrentTime(): number {
-		// GameLoopの同名メソッドとは戻り値が異なるが、 `Game.getCurrentTime()` は `Date.now()` の代替として使用されるため、整数値を返す。
-		return Math.floor(this._getCurrentTimeFunc());
-	}
-
-	raiseEvent(event: g.Event): void {
-		this.raiseEventTrigger.fire(event);
-	}
-
-	// TODO: (WIP) playlog.Event[] をとるべきか検討し対応する。
-	raiseTick(events?: g.Event[]): void {
-		if (!this.scene() || this.scene().tickGenerationMode !== g.TickGenerationMode.Manual)
-			throw g.ExceptionFactory.createAssertionError("Game#raiseTick(): tickGenerationMode for the current scene is not Manual.");
-		this.raiseTickTrigger.fire(events);
-	}
-
-	addEventFilter(filter: g.EventFilter, handleEmpty?: boolean): void {
-		this._eventFilterFuncs.addFilter(filter, handleEmpty);
-	}
-
-	removeEventFilter(filter: g.EventFilter): void {
-		this._eventFilterFuncs.removeFilter(filter);
-	}
-
-	shouldSaveSnapshot(): boolean {
-		return this.isSnapshotSaver;
-	}
-
-	// NOTE: 現状実装が `shouldSaveSnapshot()` と等価なので、簡易対応としてこの実装を用いる。
-	isActiveInstance(): boolean {
-		return this.shouldSaveSnapshot();
-	}
-
-	saveSnapshot(gameSnapshot: any, timestamp: number = this._getCurrentTimeFunc()): void {
-		if (!this.shouldSaveSnapshot())
-			return;
-		this.snapshotTrigger.fire({
-			frame: this.age,
-			timestamp,
-			data: {
-				randGenSer: this.random[0].serialize(),
-				gameSnapshot: gameSnapshot
-			}
-		});
-	}
-
 	_destroy(): void {
 		this.agePassedTrigger.destroy();
 		this.agePassedTrigger = null;
@@ -221,15 +138,7 @@ export class Game extends g.Game {
 		this.abortTrigger.destroy();
 		this.abortTrigger = null;
 		this.player = null;
-		this.raiseEventTrigger.destroy();
-		this.raiseEventTrigger = null;
-		this.raiseTickTrigger.destroy();
-		this.raiseTickTrigger = null;
-		this.snapshotTrigger.destroy();
-		this.snapshotTrigger = null;
-		this.isSnapshotSaver = false;
-		this._getCurrentTimeFunc = null;
-		this._eventFilterFuncs = null;
+		this.handlerSet = null;
 		this._notifyPassedAgeTable = null;
 		this._gameArgs = null;
 		this._globalGameArgs = null;
@@ -237,22 +146,15 @@ export class Game extends g.Game {
 	}
 
 	_restartWithSnapshot(snapshot: any): void {
-		let data = <StartPointData>snapshot.data;
-		this._eventFilterFuncs.removeFilter();
+		const data: StartPointData = snapshot.data;
 		if (data.seed != null) {
 			// 例外ケース: 第0スタートポイントでスナップショットは持っていないので特別対応
-			let randGen = new g.XorshiftRandomGenerator(data.seed);
-			this._reset({ age: snapshot.frame, randGen: randGen });
+			this._reset({ age: snapshot.frame, randSeed: data.seed });
 			this._loadAndStart({ args: this._gameArgs, globalArgs: this._globalGameArgs });
 		} else {
-			let randGen = new g.XorshiftRandomGenerator(0, data.randGenSer);
-			this._reset({ age: snapshot.frame, randGen: randGen });
+			this._reset({ age: snapshot.frame, randGenSer: data.randGenSer });
 			this._loadAndStart({ snapshot: data.gameSnapshot });
 		}
-	}
-
-	_leaveGame(): void {
-		// do nothing.
 	}
 
 	_abortGame(): void {