+++ src/GameDriver.ts
@@ -2,13 +2,14 @@
 import { Promise } from "es6-promise";
 import * as amf from "@akashic/amflow";
 import * as g from "@akashic/akashic-engine";
-import * as pdi from "@akashic/akashic-pdi";
+import * as pdi from "@akashic/pdi-types";
 import * as pl from "@akashic/playlog";
 import ExecutionMode from "./ExecutionMode";
 import LoopConfiguration from "./LoopConfiguration";
 import DriverConfiguration from "./DriverConfiguration";
 import StartPointData from "./StartPointData";
 import { Game } from "./Game";
+import { GameHandlerSet } from "./GameHandlerSet";
 import { EventBuffer } from "./EventBuffer";
 import { GameLoop } from "./GameLoop";
 import { PdiUtil } from "./PdiUtil";
@@ -245,7 +246,7 @@ export class GameDriver {
 		game.width = width;
 		game.height = height;
 		game.resized.fire({ width, height });
-		game.modified = true;
+		game.modified();
 	}
 
 	doInitialize(param: GameDriverInitializeParameterObject): Promise<void> {
@@ -412,7 +413,7 @@ export class GameDriver {
 				this._playToken = playToken;
 				this._permission = permission;
 				if (this._game) {
-					this._game.isSnapshotSaver = this._permission.writeTick;
+					this._game.handlerSet.isSnapshotSaver = this._permission.writeTick;
 				}
 				resolve();
 			});
@@ -495,12 +496,17 @@ export class GameDriver {
 				rendererCandidates: (<any>conf).renderers   // TODO: akashic-engineのGameConfigurationにrenderersの定義を加える
 			};
 			pf.setRendererRequirement(rendererRequirement);
-			var game = new Game({
+			const handlerSet = new GameHandlerSet({
+				isSnapshotSaver: this._permission.writeTick
+			});
+			const game = new Game({
+				engineModule: g,
+				handlerSet,
 				configuration: conf,
+				selfId: player.id,
 				player: player,
 				resourceFactory: pf.getResourceFactory(),
 				assetBase: param.assetBase,
-				isSnapshotSaver: this._permission.writeTick,
 				operationPluginViewInfo: (pf.getOperationPluginViewInfo ? pf.getOperationPluginViewInfo() : null),
 				gameArgs: args,
 				globalGameArgs: globalArgs
@@ -508,7 +514,7 @@ export class GameDriver {
 			var eventBuffer = new EventBuffer({ game: game, amflow: pf.amflow });
 			eventBuffer.setMode(driverConf.eventBufferMode);
 			pf.setPlatformEventHandler(eventBuffer);
-			game.setEventFilterFuncs({
+			handlerSet.setEventFilterFuncs({
 				addFilter: eventBuffer.addFilter.bind(eventBuffer),
 				removeFilter: eventBuffer.removeFilter.bind(eventBuffer)
 			});
@@ -526,14 +532,14 @@ export class GameDriver {
 			});
 
 			gameLoop.rawTargetTimeReachedTrigger.add(game._onRawTargetTimeReached, game);
-			game.setCurrentTimeFunc(gameLoop.getCurrentTime.bind(gameLoop));
-			game._reset({ age: 0, randGen: new g.XorshiftRandomGenerator(seed) });
+			handlerSet.setCurrentTimeFunc(gameLoop.getCurrentTime.bind(gameLoop));
+			game._reset({ age: 0, randSeed: seed });
 			this._updateGamePlayId(game);
 			if (this._hidden)
 				game._setMuted(true);
 
-			game.snapshotTrigger.add((startPoint: amf.StartPoint) => {
-				this._platform.amflow.putStartPoint(startPoint, (err: Error | null) => {
+			handlerSet.snapshotTrigger.add(startPoint => {
+				this._platform.amflow.putStartPoint(startPoint, err => {
 					const error = this._getCallbackError(err);
 					if (error) {
 						this.errorTrigger.fire(error);